like function added

+++ back/controllers/posts.js
@@ -9,11 +9,11 @@ exports.addPost = async (req, res, next) => {
     ) {
         return res.status(403).json({error: "Requête refusée"}) //to protect from scripts being added
     }
+    let timeOfUpload = Date.now().toString()
     const post = new Post({ // creates a new object
-        _id: Post.length,
         user: PostObject.user,
         value: PostObject.value,
-        timeOfUpload: Date.now(),
+        timeOfUpload: timeOfUpload,
         imageUrl: req.file ? `${req.protocol}://${req.get('host')}/images/${req.file.filename}` : "",
         likes: 0,
         dislikes: 0,
@@ -21,19 +21,23 @@ exports.addPost = async (req, res, next) => {
         usersDisliked: [],
     })
     post.save() //saves the new object to the DB
-    .then(() => res.status(201).json({message: 'objet enregistré'}))
+    .then(() => res.status(201).json({
+        post
+    }))
     .catch(error => res.status(400).json({error}))
 }
 
 exports.updatePost = async (req, res, next)=> {
     Post.findOne({ _id: req.params.id}) //gets the Post that will be modified from DB
     .then(post => {
-        const filename = post.imageUrl.split('/images/')[1]
-        if (fs.existsSync(`images/${filename}`) && req.file){ //if the file already exists and a file is added in the request
-            fs.unlink(`images/${filename}`, err => {if(err) { throw err}}) //deletes the file from the server
+        if(post.imageUrl !== ''){
+            const filename = post.imageUrl.split('/images/')[1]
+            if (fs.existsSync(`images/${filename}`) && req.file){ //if the file already exists and a file is added in the request
+                fs.unlink(`images/${filename}`, err => {if(err) { throw err}}) //deletes the file from the server
+            }
         }
         const PostObject = req.file ? { //if a file is added
-            ...JSON.parse(req.body.Post),
+            ...JSON.parse(req.body),
             imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}` //get the req and the infos of the file
         } : { ...req.body} //else just get the modified info from request
         if(PostObject.value.includes("<" || "javascript" || "script")
@@ -49,15 +53,17 @@ exports.updatePost = async (req, res, next)=> {
 exports.deletePost = async (req, res, next) => {
     Post.findOne({ _id: req.params.id}) //checks the DB for specific object
     .then(post => {
-        if (fs.existsSync(`images/${filename}`)){
+        if(post.imageUrl !== ''){
             const filename = post.imageUrl.split('/images/')[1]
-            fs.unlink(`images/${filename}`, (err) => {
-                if (err) throw err;
-            })
-            Post.deleteOne({ _id: req.params.id}) //deletes the object from DB
-            .then(()=> res.status(200).json({ message: 'Post supprimée!' }))
-            .catch(error => res.status(400).json({error}))
+            if (fs.existsSync(`images/${filename}`)){
+                fs.unlink(`images/${filename}`, (err) => {
+                    if (err) throw err;
+                })
+            }
         }
+        Post.deleteOne({ _id: req.params.id}) //deletes the object from DB
+        .then(()=> res.status(200).json({ message: 'Post supprimée!' }))
+        .catch(error => res.status(400).json(new Error))
     })
     .catch(error => res.status(500).json({error}))
 }
@@ -84,7 +90,7 @@ exports.likeOnePost = async (req, res, next) => {
 
         let add = (user) => { //adds the user to DB and adds a like or dislike to DB
             user.push(userId)
-            user == usersLiked? post.likes++ : post.dislikes++
+            user === usersLiked? post.likes++ : post.dislikes++
         }
         let remove = (user) => { //removes the user to DB and removes a like or dislike to DB
             let index = user.indexOf(userId)
